:bug: fix: Fix react query fetch error
